lesson3.1_fix1
